Refractor code base and also remove unnecessary import
