--- src/main/java/RITIGM/gokartproject/model/responseReceiver/common/LoginCreds.java
@@ -1,25 +1,0 @@
-package RITIGM.gokartproject.model.responseReceiver.common;
-
-public class LoginCreds {
-    private String username;
-    private String password;
-
-    public LoginCreds(String username, String password) {
-        this.username = username;
-        this.password = password;
-    }
-    public String getUsername() {
-        return username;
-    }
-    public void setUsername(String username) {
-        this.username = username;
-    }
-    public String getPassword() {
-        return password;
-    }
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    
-}
