+++ src/test/java/RITIGM/gokartproject/model/responseReceiver/CreateUIDTest.java
@@ -1,16 +1,19 @@
-package RITIGM.gokartproject.model.responseReceiver.common;
+package RITIGM.gokartproject.model.responseReceiver;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-public class NoUIDTest {
-    NoUID check;
+import RITIGM.gokartproject.model.responseReceiver.CreateUID;
+
+public class CreateUIDTest {
+
+    CreateUID check;
 
     @BeforeEach
     void test(){
-        this.check = new NoUID("1", "2", "3");
+        this.check = new CreateUID("1", "2", "3", 3);
     }
 
     @Test
@@ -21,6 +24,11 @@ public class NoUIDTest {
     @Test
     void testGetPassword() {
         assertEquals("3", check.getPassword());
+    }
+
+    @Test
+    void testGetUid() {
+        assertEquals(3, check.getUid());
     }
 
     @Test
@@ -38,6 +46,12 @@ public class NoUIDTest {
     void testSetPassword() {
         check.setPassword("727");
         assertEquals("727", check.getPassword());
+    }
+
+    @Test
+    void testSetUid() {
+        check.setUid(123);
+        assertEquals(123, check.getUid());
     }
 
     @Test
