+++ src/test/java/RITIGM/gokartproject/view/gameService/PlayerInfoServiceTest.java
@@ -12,9 +12,9 @@ import org.springframework.http.ResponseEntity;
 
 import RITIGM.gokartproject.persistence.gameService.interfaces.PlayerInfoInterface;
 import RITIGM.gokartproject.model.PlayerInfo;
-import RITIGM.gokartproject.model.responseReceiver.common.CreateUID;
-import RITIGM.gokartproject.model.responseReceiver.common.LoginCreds;
-import RITIGM.gokartproject.model.responseReceiver.common.NoUID;
+import RITIGM.gokartproject.model.responseReceiver.CreateUID;
+import RITIGM.gokartproject.model.responseReceiver.LoginCreds;
+import RITIGM.gokartproject.model.responseReceiver.NoUID;
 
 /**
  * Test for playerInfoService in view
